+++ pb/Sample5.java
@@ -1,0 +1,10 @@
+package pb;
+
+class Sample5
+{
+  public static void main(String[] args)
+  {
+    pc.Car car1 = new pc.Car();
+    car1.show();
+  }
+}
